+++ playwright.config.ts
@@ -61,7 +61,10 @@ export default defineConfig({
     /* Test against mobile viewports. */
     {
       name: 'Mobile Chrome',
-      use: { ...devices['Pixel 5'], channel: 'chrome' },
+      use: {
+        ...devices['Pixel 5'],
+        channel: 'chrome',
+      },
       // timeout: 120_000,
     },
     // {