+++ src/features/list-chapters/components/ChaptersList.tsx
@@ -1,29 +1,29 @@
 import { IonItemGroup, IonSpinner, useIonViewDidEnter } from '@ionic/react';
 import React, { useMemo, useRef, useState } from 'react';
 import { ChapterSortBy } from '../types/Chapter';
-import { useChapersList } from '../api/useChapersList';
-import { useAllJuzs } from '../api/useAllJuzs';
 import { ChapterItem } from './ChapterItem';
 import { SortedByJuz } from './SortedByJuz';
 import { FixedSizeList as List } from 'react-window';
-import DisplayError from 'components/DisplayError';
+import { Chapter } from '../types/Chapter';
+import { Juz } from '../types/Juz';
 
 interface ChapetersListProps {
   sortBy: ChapterSortBy;
   search: string;
+  chapters: Chapter[];
+  juzs: Juz[];
 }
 
 const ChaptersList: React.FC<ChapetersListProps> = ({
   sortBy = 'surah',
   search,
+  chapters,
+  juzs,
 }) => {
-  const contentRef = useRef<HTMLDivElement>(null);
+  const contentRef = useRef<HTMLIonItemGroupElement>(null);
   const [chapterListHeight, setChapterListHeight] = useState(
     window.innerHeight
   );
-
-  const { isLoading, error, data: chapterData } = useChapersList();
-  const { data: allJuzsData } = useAllJuzs();
 
   useIonViewDidEnter(() => {
     setChapterListHeight(
@@ -33,53 +33,43 @@ const ChaptersList: React.FC<ChapetersListProps> = ({
 
   const searchedChapters = useMemo(() => {
     if (search.trim()) {
-      return chapterData?.chapters?.filter((chapter) => {
+      return chapters?.filter((chapter) => {
         const chapterValues =
           `${chapter.translated_name.name}, ${chapter.id}, ${chapter.name_arabic}, ${chapter.name_simple}`.toLocaleLowerCase();
 
         return chapterValues.includes(search.toLowerCase());
       });
     }
-  }, [chapterData?.chapters, search]);
+  }, [chapters, search]);
 
   const sortedChapters = useMemo(() => {
     const sorted =
       sortBy === 'revelation-order'
-        ? chapterData?.chapters.sort(
-            (a, b) => a.revelation_order - b.revelation_order
-          )
-        : chapterData?.chapters.sort((a, b) => a.id - b.id);
+        ? chapters.sort((a, b) => a.revelation_order - b.revelation_order)
+        : chapters.sort((a, b) => a.id - b.id);
 
     return sorted;
-  }, [chapterData, sortBy]);
+  }, [chapters, sortBy]);
 
   // chapters after sorted and searched
-  const chapters = useMemo(
+  const updatedChaptersArr = useMemo(
     () => searchedChapters ?? sortedChapters ?? [],
     [searchedChapters, sortedChapters]
   );
 
   return (
-    <div className="my-3 h-[calc(100%-6.6rem)]" ref={contentRef}>
-      {/* when error appears */}
-      {error ? <DisplayError error={error} /> : null}
-
-      {/* when api is loading */}
-      {isLoading && (
-        <div className="w-full h-full grid place-items-center">
-          <IonSpinner />
-        </div>
-      )}
-
-      <IonItemGroup className="h-full">
+    <div className="h-full">
+      <IonItemGroup className="h-full" ref={contentRef}>
         {sortBy !== 'juz' ? (
           //  when succesfull data retrieve; and sortBy == 'revelation-order' or 'surah'
           <List
             height={chapterListHeight}
-            itemCount={chapters.length}
+            style={{ height: '100%' }}
+            itemCount={updatedChaptersArr.length}
             itemSize={62}
             width="100%"
-            itemData={chapters}
+            itemData={updatedChaptersArr}
+            className="ion-content-scroll-host"
           >
             {({ index, style, data }) => (
               <ChapterItem
@@ -95,8 +85,8 @@ const ChaptersList: React.FC<ChapetersListProps> = ({
         ) : (
           // when succesfull data retrieve; and sortBy == 'juzs'
           <SortedByJuz
-            chapters={searchedChapters ?? chapterData?.chapters}
-            juzs={allJuzsData?.juzs}
+            chapters={searchedChapters ?? updatedChaptersArr}
+            juzs={juzs}
           />
         )}
       </IonItemGroup>