+++ src/features/list-chapters/index.ts
@@ -4,3 +4,4 @@ export * from './types/Juz';
 export * from './api/useChapersList';
 export * from './api/useAllJuzs';
 export * from './components/ChapterItem';
+export * from './components/SortedByJuz';