+++ CHANGELOG.md
@@ -1,3 +1,55 @@
+# [1.4.0-beta.5](https://github.com/Muhammed-Rahif/Islam/compare/v1.4.0-beta.4...v1.4.0-beta.5) (2023-04-13)
+
+
+### Features
+
+* **prayer-times:** made methods dynamic ([a1aaff8](https://github.com/Muhammed-Rahif/Islam/commit/a1aaff82a1cb2bae3a5b6c1aaa1d9fbd536f2ebd))
+* **prayer-times:** prayer notifications select settings ([5f67a04](https://github.com/Muhammed-Rahif/Islam/commit/5f67a0478c8a5c885bd1e93564cf848a6fe90c47))
+
+# [1.4.0-beta.4](https://github.com/Muhammed-Rahif/Islam/compare/v1.4.0-beta.3...v1.4.0-beta.4) (2023-04-12)
+
+
+### Bug Fixes
+
+* **prayer-times:** next prayer card on complete action ([b9e1523](https://github.com/Muhammed-Rahif/Islam/commit/b9e1523e0bb75b45f1d94a427a6415fd3149aa4c))
+
+# [1.4.0-beta.3](https://github.com/Muhammed-Rahif/Islam/compare/v1.4.0-beta.2...v1.4.0-beta.3) (2023-04-11)
+
+
+### Bug Fixes
+
+* pull refresh bg color issue ([7f5b3f0](https://github.com/Muhammed-Rahif/Islam/commit/7f5b3f013dab244052f1dd3ef10772a9a29c13d2))
+* **ui:** add retry btn for 'DisplayError' ([59d3c91](https://github.com/Muhammed-Rahif/Islam/commit/59d3c9131ebf71cf97ea3a51c8a8e41df3a67e83))
+
+
+### Features
+
+* **prayer-times:** notificate user on prayer times ([4e5404a](https://github.com/Muhammed-Rahif/Islam/commit/4e5404aee6b46fed93fec14d1149e1748db8831d))
+
+# [1.4.0-beta.2](https://github.com/Muhammed-Rahif/Islam/compare/v1.4.0-beta.1...v1.4.0-beta.2) (2023-04-10)
+
+
+### Bug Fixes
+
+* signed app release ([27f11a8](https://github.com/Muhammed-Rahif/Islam/commit/27f11a811aab41a9d727e7cb785d9747bd55e72a))
+
+# [1.4.0-beta.1](https://github.com/Muhammed-Rahif/Islam/compare/v1.3.1...v1.4.0-beta.1) (2023-04-10)
+
+
+### Bug Fixes
+
+* 'NextPrayerCard' ([2d0e6de](https://github.com/Muhammed-Rahif/Islam/commit/2d0e6de9aec99af68021ddb6e7f3b69d42a93b30))
+* **performance:** memoize in list prayer times; quran search bar bg color issue ([85a02cb](https://github.com/Muhammed-Rahif/Islam/commit/85a02cb7033adb5c144835971aa5eff9eafd1c2b))
+* **ui:** ui toolbar update and improvements ([6d3d068](https://github.com/Muhammed-Rahif/Islam/commit/6d3d068c203cd3f9b84938adfad3339379625a44))
+* update tab icons ([71df5c9](https://github.com/Muhammed-Rahif/Islam/commit/71df5c9c416dc605ae1c1e8f09e2d93bb49666dc))
+
+
+### Features
+
+* **prayer-times:** add lorem ipsum of the prayer times screen ([5673c89](https://github.com/Muhammed-Rahif/Islam/commit/5673c89f002441d801e79fd9cfa4718bd78483d2))
+* **prayer-times:** add prayer times list ([77fa1c4](https://github.com/Muhammed-Rahif/Islam/commit/77fa1c4a2e66d02ed5db032ea4d10d7d0803fbca))
+* **prayer-times:** update error displaying ([59c7b0b](https://github.com/Muhammed-Rahif/Islam/commit/59c7b0b5940cec14d15efe888147d1649ef9d455))
+
 ## [1.3.1](https://github.com/Muhammed-Rahif/Islam/compare/v1.3.0...v1.3.1) (2023-04-07)